Major refactoring and add text chunk exporter
- Add a text chunk export that can take an arbitrary size and find
(complete) articles (chosen randomishly) to correspond to that size.
There are also various filters, like min/max article size and
presence/absence of a particular string (not a regex, just a string)

- Migrate to using a pyproject.toml
- Migrate to using poetry for build and release
- Add pre-commit with various best practice checks
- Make pass (with plenty of exceptions still) both pylint and black
- Make the code much more compliant to PEP8, and use snake_casing
(except for the global variable, that should be completely removed)

Also remove the "cirrus exporter" and "clean" command line apps. These
may be useful later but can be rehabilitated if required

There is plenty still to do but it is a start!